+++ Vilicus.py
@@ -140,7 +140,7 @@ def convert_to_h265(source_file_path, fail_file_path):
         output_file = base + '.mkv'
         logging.debug(output_file)
         os.rename(source_file_path, source_file_path + '.old')
-        ffmpeg.input(source_file_path + '.old').output(output_file, vcodec="libx265", crf=28, acodec="copy", **{'threads:v': '1', 'x265-params': 'aq-mode=3'}).run()
+        ffmpeg.input(source_file_path + '.old').output(output_file, vcodec='hevc_nvenc', preset='slow', global_quality=23, c='copy').run()
         ffmpeg.input(output_file).output("null", f="null").run()
         logging.info("Video validation succeeded.")
         before_file_size = os.path.getsize(source_file_path + '.old')
@@ -161,7 +161,6 @@ def convert_to_h265(source_file_path, fail_file_path):
         logging.error(f"Failed to convert file: {source_file_path}")
         with open(fail_file_path, 'w') as file:
             file.write(source_file_path)
-
 
 def main():
     global conversion_counter
